Issue #377 - use the VMUSB drivers in place rather than as copies
Import the XLM Timestamp firmware support from vmusb/daqconfig directly
instead of using a local, unmaintainable copy.
