+++ main/usb/vmusb/daqconfig/CXLMTimestamp.h
@@ -3,7 +3,9 @@
 #define CXLMTIMESTAMP_H
 
 #include <CXLM.h>
-
+#ifdef MVLC_GENERATOR
+#include <DeviceCommand.h>>
+#endif
 class CXLMTimestamp : public XLM::CXLM
 {
 
@@ -19,12 +21,33 @@ private:
 
 
 public:
+#ifdef MVLC_GENERATOR
+  virtual void onAttach(XXUSB::CConfigurableObject & configuration);
+#else
   virtual void onAttach(CReadoutModule& configuration);
+#endif
   virtual void Initialize(CVMUSB& controller);
   virtual void addReadoutList(CVMUSBReadoutList& list);
+#ifndef MVLC_GENERATOR
   virtual CReadoutHardware* clone() const; 
+#endif
 
 };
+
+#ifdef MVLC_GENERATOR
+
+// Creator command:
+
+class XLMTSCommand : public DeviceCommand {
+public:
+  XLMTSCommand(CTCLInterpreter& interp, TCLConfigParser& parser);
+  virtual ~XLMTSCommand();
+
+protected:
+  virtual CReadoutModule* createDevice(std::string name);
+};
+#endif
+
 
 
 #endif
